+++ wasmbinding/helpers/test_helpers.go
@@ -24,8 +24,16 @@ import (
 func CreateTestInput(t *testing.T) (*app.KiichainApp, sdk.Context) {
 	t.Helper()
 	chain := helpers.Setup(t)
-	ctx := chain.BaseApp.NewUncachedContext(true, tenderminttypes.Header{Height: 1, ChainID: "testing", Time: time.Now().UTC()})
+	ctx := chain.BaseApp.NewUncachedContext(true, tenderminttypes.Header{Height: 1, ChainID: "testing_1010-1", Time: time.Now().UTC()})
+	allVal, err := chain.StakingKeeper.GetAllValidators(ctx)
+	require.NoError(t, err)
 
+	// Validator consensus address
+	valConsAddr, err := allVal[0].GetConsAddr()
+	require.NoError(t, err)
+
+	// Set a final context with the proposer address for the EVM module
+	ctx = chain.BaseApp.NewUncachedContext(true, tenderminttypes.Header{Height: 1, ChainID: "testing_1010-1", Time: time.Now().UTC(), ProposerAddress: valConsAddr})
 	return chain, sdk.UnwrapSDKContext(ctx)
 }
 
@@ -86,11 +94,15 @@ func FundAccount(t *testing.T, ctx sdk.Context, app *app.KiichainApp, addr sdk.A
 // SetupCustomApp sets up a custom app for testing
 func SetupCustomApp(t *testing.T, addr sdk.AccAddress) (*app.KiichainApp, sdk.Context) {
 	t.Helper()
+
+	// Start a new app
 	app, ctx := CreateTestInput(t)
 	wasmKeeper := app.WasmKeeper
 
+	// Store the reflect code
 	storeReflectCode(t, ctx, app, addr)
 
+	// Check if all is correct with the code
 	cInfo := wasmKeeper.GetCodeInfo(ctx, 1)
 	require.NotNil(t, cInfo)