+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Update pipelines by adding codeql, codecov and changelog diff checker
 - Refactor the tokenfactory wasmbinding into its own path
 - Refactor the wasmbinding implementation to allow multiple msg and query types
+- Add EVM wasmbinding queries